Time: 4 ms (42.86%) | Memory: 72.8 MB (44.08%) - LeetSync

+++ 1803-average-waiting-time/average-waiting-time.java
@@ -1,13 +1,13 @@
 class Solution {
     public double averageWaitingTime(int[][] customers) {
         int max = 0;
-        Double sum = (double) 0;
+        long sum = 0;
         for (int arr[] : customers) {
             max = Math.max(max, arr[0]);
             sum += (arr[1] + max - arr[0]);
             max = max + arr[1];
         }
-        double len = customers.length;
-        return sum / len;
+        int len = customers.length;
+        return (double) sum / (double) len;
     }
 }
